fix: prefer WebSockets over XMPP
Previously we preferred XMPP because there was a concern that the
WebSockets API might change or go away, however since the product
line has been discontinued there is much less risk of a firmware
update breaking the WebSockets. Since the WebSocket version is
more stable, and does not have the blocking I/O issues that
XMPP does, its generally a better choice to try first.

https://codeberg.org/poezio/slixmpp/issues/3582